requirements.txt
# GUI Dependencies
PyQt5>=5.15.0
customtkinter>=5.2.0
pillow>=10.0.0

# Document Processing
pdfplumber>=0.7.6
python-docx>=0.8.11
PyMuPDF>=1.22.5

# Data Processing & AI
pandas>=2.0.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
langchain>=0.0.27
spacy>=3.6.0
faiss-cpu>=1.7.4
pyarrow>=13.0.0
openai>=1.0.0
google-generativeai>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
reportlab>=4.0.0

# Authentication (optional - for login feature)
pyrebase>=3.0.0

# Additional dependencies
requests>=2.32.0  # Python 3.14 compatibility
scikit-learn>=1.0.0  # For LDA and vectorization in reports

# Evidence Processing & OCR
pytesseract>=0.3.10  # OCR for image text extraction
opencv-python>=4.8.0  # Image processing
Pillow>=10.0.0  # Already included, but ensuring for image handling